Criado popups para buttons, confirms de acoes com sweetalert2

+++ frontend/src/components/list-user/ListUserScript.js
@@ -18,29 +18,45 @@ export default {
 
     async removeUser(id) {
       // Deleta dps lista novamente
-      await UserDataServices.deleteUser(id);
-      await this.listAllUsers();
-    //   this.$swal({
-    //     title: 'Deletar user?',
-    //     text: 'Cuidado! Este user sera deletado!',
-    //     icon: 'Perigo',
-    //     showConfirmButton: true,
-    //     allowOutsideClick: false,
-    //     allowEnterKey: true,
-    //     allowEscapeKey: false,
-    //     showCancelButton: true,
-    //     confirmButtonColor: '#3085d6',
-    //     cancelButtonColor: '#d33',
-    //     confirmButtonText: 'Sim! deletar!!',
-    //   }).then(async (result) => {
-    //     if (result.value) {
-    //       await UserDataServices.deleteUser(id);
-    //       this.$swal('Deleted', 'User deletado!', 'success');
-    //       this.listAllUsers();
-    //     } else {
-    //       this.$swal('Cancelled', 'Cancel deletion', 'info');
-    //     }
-    //   });
+      this.$swal({
+        title: 'Deletar user',
+        text: 'Cuidado! Este user sera deletado!',
+        icon: 'warning',
+        showConfirmButton: true,
+        allowOutsideClick: false,
+        allowEnterKey: true,
+        allowEscapeKey: false,
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Sim, deletar!',
+        cancelButtonText: 'Cancelar',
+      }).then(async (result) => {
+        // console.log(result.isConfirmed);
+        if (result.isConfirmed) {
+          await UserDataServices.deleteUser(id);
+          this.$swal({
+            title: 'User deletado',
+            icon: 'success',
+            toast: true,
+            showConfirmButton: false,
+            timer: 1500,
+            timerProgressBar: true,
+            position: 'top-end',
+          });
+          await this.listAllUsers();
+        } else {
+          this.$swal({
+            title: 'User não deletado',
+            icon: 'info',
+            toast: true,
+            showConfirmButton: false,
+            timer: 1500,
+            timerProgressBar: true,
+            position: 'top-end',
+          });
+        }
+      });
     },
 
     mudarPaginaUpdate(paramId) {
